+++ src/Tracing/GuzzleTracingMiddleware.php
@@ -18,20 +18,20 @@ final class GuzzleTracingMiddleware
         return function (callable $handler) use ($hub): \Closure {
             return function (RequestInterface $request, array $options) use ($hub, $handler) {
                 $hub = $hub ?? SentrySdk::getCurrentHub();
-                $transaction = $hub->getTransaction();
-                $span = null;
+                $span = $hub->getSpan();
+                $childSpan = null;
 
-                if (null !== $transaction) {
+                if (null !== $span) {
                     $spanContext = new SpanContext();
                     $spanContext->setOp('http.guzzle');
                     $spanContext->setDescription($request->getMethod() . ' ' . $request->getUri());
 
-                    $span = $transaction->startChild($spanContext);
+                    $childSpan = $span->startChild($spanContext);
                 }
 
-                $handlerPromiseCallback = static function ($responseOrException) use ($span) {
-                    if (null !== $span) {
-                        $span->finish();
+                $handlerPromiseCallback = static function ($responseOrException) use ($childSpan) {
+                    if (null !== $childSpan) {
+                        $childSpan->finish();
                     }
 
                     if ($responseOrException instanceof \Throwable) {