+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fix: Use Span on Scope instead of Transaction for GuzzleMiddleware (#1099)
+
 ## 3.0.0 (2020-09-28)
 
 **Tracing API**