+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - fix: Use Span on Scope instead of Transaction for GuzzleMiddleware (#1099)
+- Add setter for value on the `ExceptionDataBag` (#1100)
 
 ## 3.0.0 (2020-09-28)