+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - fix: Use Span on Scope instead of Transaction for GuzzleMiddleware (#1099)
-- 
+- feat: Expose setDsn in Options (#1105)
 
 ## 3.0.0 (2020-09-28)